requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
tensorflow==2.6.0
Keras==2.6.0
numpy==1.19.5
Pillow==9.0.1
protobuf==3.20.1
